Fix tests for SimpleOCR
